Allow Zxcvbn to be used asynchronously
Provides a second `async-password-strength` shim which loads zxcvbn on
demand. This allows the very heavy zxcvbn to be loaded only when it is
needed.

Adds a build option bundleZxcvbn which defaults to true to maintain the
current behavior. When this option is set to false zxcvbn will not be
bundled into the apps `vendor.js` file.