Add passwordStrength service and deprecate the shim
The password strength service gives us the ability to use Zxcvbn
asynchronously and load it only when needed saving significant and often
unnecessary space in our vendor.js file.

Adds a build option bundleZxcvbn which defaults to true to maintain the
current behavior. When this option is set to false zxcvbn will not be
bundled into the apps vendor.js file.